Fuel now drains when engine is used and increases as cargo fills up. Oxygen also drains but not when the player is selecting a game event.

+++ Resources.h
@@ -3,8 +3,8 @@
 const unsigned int numberOfResources = 10;
 struct Resources {
 	int health;
-	int fuel;
-	int oxygen;
+	float fuel;
+	float oxygen;
 	int titanium;
 	int scrapMetal;
 	int nanotech;
